Putting arm state enums into separate files
