+++ ROBOT_CODE/RobotBase/src/main/java/frc/robot/state/arm/ArmInput.java
@@ -1,0 +1,14 @@
+package frc.robot.state.arm;
+
+import frc.robot.state.*;
+
+/*
+ * Defines valid inputs for state transitions
+ */
+public enum ArmInput implements Input {
+    EXTEND, RETRACT, INTAKE, RELEASE, RECOVER, INTERRUPT, SUCCESS, FAILED;
+
+    public String getDescription() {
+        return "ArmStateInput: " + this.toString();
+    }
+}
