+++ ROBOT_CODE/RobotBase/src/test/java/frc/robot/state/arm/ArmStateMachineTest.java
@@ -9,9 +9,9 @@ import org.junit.jupiter.api.Test;
 import frc.robot.state.*;
 import frc.robot.state.StateMachine.Status;
 import frc.robot.state.arm.ArmStateMachine;
-import frc.robot.state.arm.ArmStateMachine.ArmInput;
-import frc.robot.state.arm.ArmStateMachine.ArmState;
-import frc.robot.state.arm.ArmStateMachine.ArmSequence;
+import frc.robot.state.arm.ArmInput;
+import frc.robot.state.arm.ArmState;
+import frc.robot.state.arm.ArmSequence;
 import frc.robot.Constants.StateConstants;
 import frc.robot.Constants.StateConstants.StateMachineWaitCondition;
 import frc.robot.testsupport.mock.MockStateSubsystem;
